MOBILEAPPS-511 Creating MS Office documents

+++ ContentApp/BusinessLayer/Action Menu/ViewModels/NodeActionsViewModel.swift
@@ -28,20 +28,20 @@ let kSaveToCameraRollAction = "com.apple.UIKit.activity.SaveToCameraRoll"
 
 protocol NodeActionsViewModelDelegate: class {
     func nodeActionFinished(with action: ActionMenu?,
-                            node: ListNode,
+                            node: ListNode?,
                             error: Error?)
 }
 
 class NodeActionsViewModel {
     private var action: ActionMenu?
-    private var node: ListNode
+    private var node: ListNode?
     private var actionFinishedHandler: ActionFinishedCompletionHandler?
     private var coordinatorServices: CoordinatorServices?
     weak var delegate: NodeActionsViewModelDelegate?
 
     // MARK: Init
 
-    init(node: ListNode,
+    init(node: ListNode? = nil,
          delegate: NodeActionsViewModelDelegate?,
          coordinatorServices: CoordinatorServices?) {
         self.node = node
@@ -84,16 +84,17 @@ class NodeActionsViewModel {
             case .download :
                 sSelf.requestDownload()
             default:
-                DispatchQueue.main.async {
+                DispatchQueue.main.asyncAfter(deadline: .now() + 1.0, execute: {
                     sSelf.delegate?.nodeActionFinished(with: sSelf.action,
                                                        node: sSelf.node,
                                                        error: nil)
-                }
+                })
             }
         })
     }
 
     private func requestAddToFavorites() {
+        guard let node = self.node else { return }
         let jsonGuid = JSONValue(dictionaryLiteral: ("guid", JSONValue(stringLiteral: node.guid)))
         let jsonFolder = JSONValue(dictionaryLiteral: (node.kind.rawValue, jsonGuid))
         let jsonBody = FavoriteBodyCreate(target: jsonFolder)
@@ -102,10 +103,10 @@ class NodeActionsViewModel {
                                     favoriteBodyCreate: jsonBody) { [weak self] (_, error) in
             guard let sSelf = self else { return }
             if error == nil {
-                sSelf.node.favorite = true
+                sSelf.node?.favorite = true
                 sSelf.action?.type = .removeFavorite
                 sSelf.action?.title = LocalizationConstants.ActionMenu.removeFavorite
-                let favouriteEvent = FavouriteEvent(node: sSelf.node, eventType: .addToFavourite)
+                let favouriteEvent = FavouriteEvent(node: node, eventType: .addToFavourite)
 
                 let eventBusService = sSelf.coordinatorServices?.eventBusService
                 eventBusService?.publish(event: favouriteEvent, on: .mainQueue)
@@ -115,14 +116,15 @@ class NodeActionsViewModel {
     }
 
     private func requestRemoveFromFavorites() {
+        guard let node = self.node else { return }
         FavoritesAPI.deleteFavorite(personId: kAPIPathMe,
                                     favoriteId: node.guid) { [weak self] (_, error) in
             guard let sSelf = self else { return }
             if error == nil {
-                sSelf.node.favorite = false
+                sSelf.node?.favorite = false
                 sSelf.action?.type = .addFavorite
                 sSelf.action?.title = LocalizationConstants.ActionMenu.addFavorite
-                let favouriteEvent = FavouriteEvent(node: sSelf.node,
+                let favouriteEvent = FavouriteEvent(node: node,
                                                     eventType: .removeFromFavourites)
 
                 let eventBusService = sSelf.coordinatorServices?.eventBusService
@@ -133,12 +135,13 @@ class NodeActionsViewModel {
     }
 
     private func requestMoveToTrash() {
+        guard let node = self.node else { return }
         if node.kind == .site {
             SitesAPI.deleteSite(siteId: node.siteID) { [weak self] (_, error) in
                 guard let sSelf = self else { return }
                 if error == nil {
-                    sSelf.node.trashed = true
-                    let moveEvent = MoveEvent(node: sSelf.node, eventType: .moveToTrash)
+                    sSelf.node?.trashed = true
+                    let moveEvent = MoveEvent(node: node, eventType: .moveToTrash)
 
                     let eventBusService = sSelf.coordinatorServices?.eventBusService
                     eventBusService?.publish(event: moveEvent, on: .mainQueue)
@@ -149,8 +152,8 @@ class NodeActionsViewModel {
             NodesAPI.deleteNode(nodeId: node.guid) { [weak self] (_, error) in
                 guard let sSelf = self else { return }
                 if error == nil {
-                    sSelf.node.trashed = true
-                    let moveEvent = MoveEvent(node: sSelf.node, eventType: .moveToTrash)
+                    sSelf.node?.trashed = true
+                    let moveEvent = MoveEvent(node: node, eventType: .moveToTrash)
 
                     let eventBusService = sSelf.coordinatorServices?.eventBusService
                     eventBusService?.publish(event: moveEvent, on: .mainQueue)
@@ -161,11 +164,12 @@ class NodeActionsViewModel {
     }
 
     private func requestRestoreFromTrash() {
+        guard let node = self.node else { return }
         TrashcanAPI.restoreDeletedNode(nodeId: node.guid) { [weak self] (_, error) in
             guard let sSelf = self else { return }
             if error == nil {
-                sSelf.node.trashed = false
-                let moveEvent = MoveEvent(node: sSelf.node, eventType: .restore)
+                sSelf.node?.trashed = false
+                let moveEvent = MoveEvent(node: node, eventType: .restore)
 
                 let eventBusService = sSelf.coordinatorServices?.eventBusService
                 eventBusService?.publish(event: moveEvent, on: .mainQueue)
@@ -175,6 +179,7 @@ class NodeActionsViewModel {
     }
 
     private func requestPermanentlyDelete() {
+        guard let node = self.node else { return }
         let title = LocalizationConstants.NodeActionsDialog.deleteTitle
         let message = String(format: LocalizationConstants.NodeActionsDialog.deleteMessage,
                              node.title)
@@ -183,9 +188,9 @@ class NodeActionsViewModel {
         let deleteAction = MDCAlertAction(title: LocalizationConstants.Buttons.delete) { [weak self] _ in
             guard let sSelf = self else { return }
 
-            TrashcanAPI.deleteDeletedNode(nodeId: sSelf.node.guid) { (_, error) in
+            TrashcanAPI.deleteDeletedNode(nodeId: node.guid) { (_, error) in
                 if error == nil {
-                    let moveEvent = MoveEvent(node: sSelf.node, eventType: .permanentlyDelete)
+                    let moveEvent = MoveEvent(node: node, eventType: .permanentlyDelete)
 
                     let eventBusService = sSelf.coordinatorServices?.eventBusService
                     eventBusService?.publish(event: moveEvent, on: .mainQueue)
@@ -252,7 +257,7 @@ class NodeActionsViewModel {
             let themingService = coordinatorServices?.themingService
             downloadDialogView.messageLabel.text =
                 String(format: LocalizationConstants.NodeActionsDialog.downloadMessage,
-                       node.title)
+                       node?.title ?? "")
             downloadDialogView.activityIndicator.startAnimating()
             downloadDialogView.applyTheme(themingService?.activeTheme)
 
@@ -276,7 +281,8 @@ class NodeActionsViewModel {
     }
 
     private func downloadNodeContent(completionHandler: @escaping (URL?, APIError?) -> Void) -> DownloadRequest? {
-        let requestBuilder = NodesAPI.getNodeContentWithRequestBuilder(nodeId: self.node.guid)
+        guard let node = self.node else { return nil}
+        let requestBuilder = NodesAPI.getNodeContentWithRequestBuilder(nodeId: node.guid)
         let downloadURL = URL(string: requestBuilder.URLString)
 
         var documentsURL = FileManager.default.urls(for: .documentDirectory,
